Fix snowflakeToThreadChannel
Very slow, when testing it took arroud 3,000 ms to make a loop to get one thread.

A better approach is to use a O(log n) loop that will take less time than a O(n) loop.
We can't use this in production. (I mean, this implementation in prod (see TODO below)).
An interaction without acknowledgment give a maximum of three seconds to respond to it. With a > 3s performances, it's impossible to respond the first time. That's why I used caching, but still, we need to fix this.
That's a "hacky" way to fix (and still, it don't work the first time).

TODO: Need to re-implements `@snowflakeToThreadChannel`